OPPE 2 Mock Question 1
Papers is a list of tuples that contains information about the number of papers published by a researcher during his career. Each element is of the form (year,num): the researcher has published (num) papers in the (year) year.

Write a function named good_years hat accepts (papers) as argument and returns the list of years in which the researcher has published the maximum number of papers.